% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_shp.R
\name{load_shp}
\alias{load_shp}
\title{Loading the shape files}
\usage{
load_shp(zip_dim, quality = "high")
}
\arguments{
\item{zip_dim}{integer if zip code dimension has 1, 2, 3 or 5 digits}

\item{quality}{character string either "high" (default) or "low"}
}
\value{
tibble with geo location, population and zip code
}
\description{
Loading shape files with its polygon location, zip codes
and population.
}
